visual_analyse/requirements.txt
# opengait required
pytorch >= 1.10
torchvision
pyyaml
tensorboard
opencv-python
tqdm
py7zr
kornia
einops

# visual analysis required
rich
fire